--- src/main/java/es/in2/wallet/api/broker/properties/BrokerPathProperties.java
@@ -1,9 +1,0 @@
-package es.in2.wallet.api.broker.properties;
-
-/**
- * NGSI-LD Paths
- *
- * @param entities      - entities path
- */
-public record BrokerPathProperties(String entities) {
-}
