--- src/main/java/es/in2/wallet/api/broker/properties/BrokerProperties.java
@@ -1,30 +1,0 @@
-package es.in2.wallet.api.broker.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.NestedConfigurationProperty;
-import org.springframework.boot.context.properties.bind.ConstructorBinding;
-
-import java.util.Optional;
-
-/**
- * Configuration intended to connect the NGSI-LD ContextBroker
- *
- * @param provider       - context broker provider
- * @param externalDomain - domain that the broker is externally available. Used for the hashlink.
- * @param internalDomain - internal address of the broker, used to connect from within the connector
- * @param paths          - ngis-ld paths to be used when connecting the broker
- */
-@ConfigurationProperties(prefix = "broker")
-public record BrokerProperties(String provider, String externalDomain, String internalDomain,
-                               @NestedConfigurationProperty BrokerPathProperties paths) {
-
-    @ConstructorBinding
-    public BrokerProperties(String provider, String externalDomain, String internalDomain, BrokerPathProperties paths) {
-        this.provider = provider;
-        this.externalDomain = externalDomain;
-        this.internalDomain = internalDomain;
-        this.paths = Optional.ofNullable(paths).orElse(new BrokerPathProperties(null));
-    }
-
-}
-
