+++ src/main/java/es/in2/wallet/api/crypto/service/DidKeyGeneratorService.java
@@ -2,9 +2,9 @@ package es.in2.wallet.api.crypto.service;
 
 import reactor.core.publisher.Mono;
 
-import java.io.IOException;
 import java.security.KeyPair;
 
 public interface DidKeyGeneratorService {
-    Mono<String> generateDidFromKeyPair(KeyPair keyPair);
+    Mono<String> generateDidKeyJwkJcsPubWithFromKeyPair(KeyPair keyPair);
+    Mono<String> generateDidKeyFromKeyPair(KeyPair keyPair);
 }
