+++ src/main/java/es/in2/wallet/api/crypto/service/impl/KeyGenerationServiceImpl.java
@@ -1,6 +1,7 @@
 package es.in2.wallet.api.crypto.service.impl;
 
 import es.in2.wallet.api.crypto.service.KeyGenerationService;
+import es.in2.wallet.api.exception.KeyPairGenerationError;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
@@ -13,7 +14,7 @@ import java.security.spec.ECGenParameterSpec;
 @Service
 @Slf4j
 public class KeyGenerationServiceImpl implements KeyGenerationService {
-    public Mono<KeyPair> generateECKeyPair() {
+    public Mono<KeyPair> generateES256r1ECKeyPair() {
         return Mono.fromCallable(() -> {
             try {
                 KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("ECDSA", "BC");
@@ -21,7 +22,7 @@ public class KeyGenerationServiceImpl implements KeyGenerationService {
                 keyPairGenerator.initialize(ecSpec, new SecureRandom());
                 return keyPairGenerator.generateKeyPair();
             } catch (Exception e) {
-                throw new RuntimeException("Error generating EC key pair", e);
+                throw new KeyPairGenerationError("Error generating EC key pair: " + e);
             }
         });
     }
