+++ src/main/java/es/in2/wallet/api/exception/KeyPairGenerationError.java
@@ -1,0 +1,7 @@
+package es.in2.wallet.api.exception;
+
+public class KeyPairGenerationError extends RuntimeException {
+    public KeyPairGenerationError(String message) {
+        super(message);
+    }
+}
