--- src/main/java/es/in2/wallet/api/vault/domain/VaultSecretData.java
@@ -1,12 +1,0 @@
-package es.in2.wallet.api.vault.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record VaultSecretData(
-        @JsonProperty("privateKey") String privateKey
-) {
-}
