--- src/main/java/es/in2/wallet/api/vault/properties/SecretProvider.java
@@ -1,3 +1,0 @@
-package es.in2.wallet.api.vault.properties;
-public record SecretProvider(String name) {
-}
