--- src/main/java/es/in2/wallet/api/vault/properties/azure/Secret.java
@@ -1,4 +1,0 @@
-package es.in2.wallet.api.vault.properties.azure;
-
-public record Secret(String endpoint) {
-}
