--- src/main/java/es/in2/wallet/api/vault/service/GenericVaultService.java
@@ -1,11 +1,0 @@
-package es.in2.wallet.api.vault.service;
-
-import reactor.core.publisher.Mono;
-
-public interface GenericVaultService {
-    Mono<Void> saveSecret(String key, String secret);
-
-    Mono<String> getSecretByKey(String key);
-
-    Mono<Void> deleteSecretByKey(String key);
-}
