+++ src/main/java/es/in2/wallet/broker/properties/BrokerPathProperties.java
@@ -1,0 +1,9 @@
+package es.in2.wallet.broker.properties;
+
+/**
+ * NGSI-LD Paths
+ *
+ * @param entities      - entities path
+ */
+public record BrokerPathProperties(String entities) {
+}
