+++ src/main/java/es/in2/wallet/vault/domain/VaultSecretData.java
@@ -1,0 +1,12 @@
+package es.in2.wallet.vault.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+
+@Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record VaultSecretData(
+        @JsonProperty("privateKey") String privateKey
+) {
+}
