+++ src/main/java/es/in2/wallet/vault/properties/SecretProvider.java
@@ -1,0 +1,3 @@
+package es.in2.wallet.vault.properties;
+public record SecretProvider(String name) {
+}
