+++ src/main/java/es/in2/wallet/vault/properties/azure/Secret.java
@@ -1,0 +1,4 @@
+package es.in2.wallet.vault.properties.azure;
+
+public record Secret(String endpoint) {
+}
